Fix memory leak
So, it turns out that we do need to call sd_bus_message_unref(reply) on
the reply message we get back from sd_bus_call_method(). However, doing
so (assuming we're the last/only holder of a ref to reply) causes
libsystemd to call close() on any file descriptors *inside* the message,
which releases the inhibitor lock, which is not what we want!

This behaviour of sd_bus_message_unref() is... surprising. Anyway, work
around it by treating the reply message itself as our "lock" (rather
than just the fd), and rely on unref'ing it as the way to release the
lock.