Delay suspend by 1 second
(Hypothesis) There is a race between:

1) the X server committing the result of "xscreensaver-command -lock" to
the hardware.

and

2) systemd suspending the system.

This results in, on occasion, the "activate-deactivate" dance
effectively happening *after resume* rather than *before suspend*.

Delaying (2) by one second by doing a sleep(1) after a successful
"xscreensaver-command -lock" is a tentative workaround.

If there were a way to sync on the completion of (1) instead, that would
be a proper solution, but unclear if there is a way to do that [all the
way down to the hardware] at all.